Bgfx pointer not move since Go1.6 modification
Prior to Go1.6, pointer to pointer was used which allow C iconv
to change the value of pointer to array directly in Go func.
However, after Go1.6 bridge_iconv is used and since then causes
C iconv not able to directly change the value of pointer to
array. DoWrite requires an updated address to inbuf array for
each loop of C bridge_iconv call. C bridge_iconv was called with
the same address of first element of array and caused it to
convert the same byte content again and again if the size of
the inbuf larger than twice the size supplied of outbuf (Conv
method). This commit solved the bug with every call to
bridge_iconv, updated address of next first unconverted inbuf
element will be used.